requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.5
django-ckeditor==5.5.0
django-js-asset==1.1.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
Pillow==5.2.0
pytz==2018.4
requests==2.19.1
uritemplate==3.0.0
urllib3==1.23
wincertstore==0.2
